Fixed a formatting issue with the hotbar due to a library change

+++ src/main/java/com/thenexusreborn/survivalgames/game/GameActionBar.java
@@ -50,7 +50,7 @@ public class GameActionBar implements IActionBar {
         }
 
         if (game.getState() == GameState.WARMUP) {
-            return "&fGAME BEGINS IN &c&l" + (TimeUnit.MILLISECONDS.toSeconds(game.getTimer().getTime()) + 1) + " SECONDS&f...";
+            return "&fGAME BEGINS IN &c&l" + ((int) TimeUnit.MILLISECONDS.toSeconds(game.getTimer().getTime()) + 1) + " SECONDS&f...";
         }
 
         if (game.getState() == GameState.WARMUP_DONE) {
@@ -114,7 +114,7 @@ public class GameActionBar implements IActionBar {
             if (game.getTimer() == null) {
                 return "&aArchiving game...";
             }
-            return "&d&lNEXUS&7 - &eNext game starting... &7(" + (TimeUnit.MILLISECONDS.toSeconds(game.getTimer().getTime()) + 1) + "s)";
+            return "&d&lNEXUS&7 - &eNext game starting... &7(" + ((int) TimeUnit.MILLISECONDS.toSeconds(game.getTimer().getTime()) + 1) + "s)";
         }
 
         return "";
